refactor: move hidden field test into its own file
